REGRESSION(C stack work): stack traces no longer work in CrashTracer, lldb, and other tools https://bugs.webkit.org/show_bug.cgi?id=126764
Reviewed by Geoffrey Garen.

Updated callToJavaScript and cllToNativeFunction to properly replicate the caller's
return PC and frame pointer in the sentinel frame.  For X86-64, added .cfi_
directives to create eh_frame info for all LLInt symbols so that the various
unwinding code understands that we are using a separate JS stack referenced
by BP and at what offsets in that frame the prior PC (register 16) and prior
BP (register 6) can be found.  These two changes are sufficient for stack tracing
to work for Mac OSX.

* llint/LowLevelInterpreter.cpp:
* llint/LowLevelInterpreter64.asm:


Identifier: 144695@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@161686 268f45cc-cd09-0410-ab3c-d52691b4dbfc