EWS should provide better information to Dashboard via JSON https://bugs.webkit.org/show_bug.cgi?id=127265
Reviewed by Ryosuke Niwa.

* QueueStatusServer/app.yaml: Updated version.

* QueueStatusServer/main.py:
* QueueStatusServer/handlers/queuelengthjson.py:
Added a way to cheaply get only the queue length, which is all the dashboard
needs unless a popover is opened.

* QueueStatusServer/handlers/queuestatusjson.py: Updated to return more information
that we will need. Changed some field names to be more meaningful. Fixed bot listing
to understand that a bot can be reprurposed and used in a different queue.



Identifier: 145283@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@162358 268f45cc-cd09-0410-ab3c-d52691b4dbfc