[GTK] Remove null check for WebPopupMenuProxy::Client in WebPopupMenuProxyGtk::showPopupMenu https://bugs.webkit.org/show_bug.cgi?id=107036
Reviewed by Andreas Kling.

* UIProcess/gtk/WebPopupMenuProxyGtk.cpp:
(WebKit::WebPopupMenuProxyGtk::showPopupMenu): Don't perform an early return
if the client is not available - it always should be. The client is set to null
only when the proxy is invalidated. This happens immediately after this function
returns, either with the popup being forced to close or being closed by the user
with possibly an item being chosen. Rather than the early return, use an ASSERT.


Identifier: 145633@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@162768 268f45cc-cd09-0410-ab3c-d52691b4dbfc