clip-path swaps bottom radii for the inset shape https://bugs.webkit.org/show_bug.cgi?id=127859
Reviewed by Dirk Schulze.

Source/WebCore:

The bottom right and bottom left radii were passed in the wrong order,
causing the wrong corner to be clipped in the resultant rendering.

Test: css3/masking/clip-path-inset-corners.html

* rendering/style/BasicShapes.cpp:
(WebCore::BasicShapeInset::path): Change the parameter order to be
correct.

LayoutTests:

Test that all the inset radii are applied to the correct corners.

* css3/masking/clip-path-inset-corners-expected.html: Added.
* css3/masking/clip-path-inset-corners.html: Added.


Identifier: 145872@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@163044 268f45cc-cd09-0410-ab3c-d52691b4dbfc