WebKit2 View Gestures (Swipe): Crashes swiping around a Google Docs spreadsheet https://bugs.webkit.org/show_bug.cgi?id=128013 <rdar://problem/15959481>
Reviewed by Anders Carlsson.

* UIProcess/mac/ViewGestureController.mm:
(WebKit::ViewGestureController::wheelEventWasNotHandledByWebCore):
We should never (currently) start a swipe from a non-begin-phase wheel event.
We already check this in handleScrollWheelEvent, but can still get
non-begin-phase events in wheelEventWasNotHandled.

We only start "pending" swipes with Begin events, so we should never make them
non-pending with anything but a Begin event (we'll get it when it bubbles back up,
unhandled, from WebCore). In addition, AppKit throws an exception when we try
to start a swipe from end-phase events.


Identifier: 146013@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@163192 268f45cc-cd09-0410-ab3c-d52691b4dbfc