Text interaction assistant is not deleted when dismissing the keyboard on iPad. https://bugs.webkit.org/show_bug.cgi?id=128034 <rdar://problem/15915695>
Reviewed by Dan Bernstein.

In iPhone mode, tapping on 'Done' in the accessory view
calls the delegate accessoryDone where we correctly blur
the assisted node and tear down the text interaction assistant.
The code path for the keyboard in iPad mode is different,
all we get is resignFirstResponder.
The fix is just to move the blurring in resignFirstResponder
and it works for both iPhone and iPad.

* UIProcess/API/ios/WKInteractionView.mm:
(-[WKInteractionView resignFirstResponder]):
(-[WKInteractionView accessoryDone]):



Identifier: 146068@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@163250 268f45cc-cd09-0410-ab3c-d52691b4dbfc