REGRESSION (r159609): Images are corrupted when hovering over buttons @ github.com https://bugs.webkit.org/show_bug.cgi?id=127729
Reviewed by Antti Koivisto.

Source/WebCore:

When clipping a rect, the RenderLayer would not properly save the context when
the clipping rect is the same as the paint rect and the clipping rect
has radius.

Test: fast/regions/repaint/hover-border-radius.html

* rendering/RenderLayer.cpp:
(WebCore::RenderLayer::clipToRect):
(WebCore::RenderLayer::restoreClip):

LayoutTests:

Added test for the correct repainting of buttons with border radius when hovering.

* fast/regions/repaint/hover-border-radius-expected.txt: Added.
* fast/regions/repaint/hover-border-radius.html: Added.


Identifier: 146189@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@163382 268f45cc-cd09-0410-ab3c-d52691b4dbfc